fixed func input
