+++ Makefile
@@ -5,6 +5,9 @@ SHELL:=/usr/bin/env bash
 lint:
 	flake8 piper
 
+unit-services:
+	pytest -vs tests/services/chat_gpt.py
+
 unit:
 	# pytest -vs tests/import_test.py
 	# pytest -vs tests/base_executor_test.py
@@ -12,6 +15,7 @@ unit:
 	# pytest -vs tests/envs_test.py::TestCompose
 	# pytest -vs tests/envs_test.py::TestVenv
 	pytest -vs tests/clip_test.py
+
 
 package:
 	pip check
