+++ piper/base/executors/__init__.py
@@ -1,3 +1,4 @@
 from piper.base.executors._base_executor import BaseExecutor
 from piper.base.executors.http import HTTPExecutor
-from piper.base.executors.fastapi import FastAPIExecutor, FastAPITesseractExecutor
+from piper.base.executors.fastapi import FastAPIExecutor
+from piper.base.executors.tesseract import FastAPITesseractExecutor
